[REFACTOR] Update API to introduce service interfaces
Refactors the addon to introduce a concept of service interfaces, along
with adding `serviceOverrides` directly to the app definition. This
makes overriding a service more restrictive, but simplifies the overall
rules for how overriding works. It also ensures that users can trust
their tooling, since they'll see whenever they click on a service if it
is a concrete definition, or an interface.